+++ packages/frontend/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.7](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.7-alpha.0...@evm-debuger/frontend@0.2.7) (2023-05-18)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
+## [0.2.7-alpha.0](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.6...@evm-debuger/frontend@0.2.7-alpha.0) (2023-05-18)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
+## [0.2.6](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.6-alpha.0...@evm-debuger/frontend@0.2.6) (2023-05-17)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
+## [0.2.6-alpha.0](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.5...@evm-debuger/frontend@0.2.6-alpha.0) (2023-05-17)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
+## [0.2.5](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.4...@evm-debuger/frontend@0.2.5) (2023-05-17)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
+## [0.2.4](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.4-alpha.0...@evm-debuger/frontend@0.2.4) (2023-05-17)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
+## [0.2.4-alpha.0](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.3...@evm-debuger/frontend@0.2.4-alpha.0) (2023-05-17)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
+## [0.2.3](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.3-alpha.0...@evm-debuger/frontend@0.2.3) (2023-05-17)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
+## [0.2.3-alpha.0](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.2...@evm-debuger/frontend@0.2.3-alpha.0) (2023-05-17)
+
+**Note:** Version bump only for package @evm-debuger/frontend
+
 ## [0.2.2](https://github.com/rumblefishdev/evm-debuger/compare/@evm-debuger/frontend@0.2.2-alpha.0...@evm-debuger/frontend@0.2.2) (2023-05-04)
 
 **Note:** Version bump only for package @evm-debuger/frontend